third_party/WebKit/LayoutTests/media/controls-cast-button-narrow.html

<!DOCTYPE html>
<title>media controls cast button</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<video width="100" height="200" controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = findMediaFile("video", "content/test");

    video.onloadedmetadata = t.step_func(function() {
        // Pretend we have a cast device
        internals.mediaPlayerRemoteRouteAvailabilityChanged(video, true);

        // Video should not have a cast button since the video is too narrow
        assert_false(isVisible(castButton(video)), "button should not be visible");

        // It should not have an overlay cast button too. Instead, the button
        // should appear in the overflow menu. Tested somewhere else.
        assert_false(isVisible(overlayCastButton(video)), "button should not be visible");

        // Increasing video width should show the cast button.
        video.width = 500;
        testRunner.layoutAndPaintAsyncThen(t.step_func(function() {
            assert_true(isVisible(castButton(video)), "button should be visible");

            // Decreasing video width should hide the cast button again.
            video.width = 100;
            testRunner.layoutAndPaintAsyncThen(t.step_func(function() {
                assert_false(isVisible(castButton(video)), "button should not be visible");

                // If the cast device is lost, we still don't show the cast
                // button when the video width increases.
                internals.mediaPlayerRemoteRouteAvailabilityChanged(video, false);
                video.width = 500;
                testRunner.layoutAndPaintAsyncThen(t.step_func_done(function() {
                    assert_false(isVisible(castButton(video)), "button should not be visible");
                }));
            }));
        }));
    });

    function isVisible(button) {
        var computedStyle = getComputedStyle(button);
        return computedStyle.display !== "none" &&
               computedStyle.display !== "hidden" &&
               computedStyle.visibility === "visible";
    }
});
</script>
